Util to GridStack.createWidgetDivs() move
* fix #2959
*  `Util.createWidgetDivs()` has moved to `GridStack.createWidgetDivs()` to remove circular dependencies
